==> notebooks/00-explore-data.ipynb <==
import sys

import datasets
import pandas as pd
import plotly.express as px
from datasets import load_dataset
from pydantic import BaseModel

datasets.logging.set_verbosity_error()

sys.path.insert(0, "..")

from src.data_prep_utils import handle_dtypes, parse_dt

class Args(BaseModel):
    hf_dataset_path: str = "McAuley-Lab/Amazon-Reviews-2023"
    report_sample_num_rows: int = 10000
    random_seed: int = 41


args = Args()

print(args.model_dump_json())

metadata = load_dataset(
    args.hf_dataset_path,
    name="raw_meta_Toys_and_Games",
    split="full",
    trust_remote_code=True,
)
metadata[0]

dataset = load_dataset(
    args.hf_dataset_path, name="5core_timestamp_Toys_and_Games", trust_remote_code=True
)
train_raw_df = dataset["train"].to_pandas()
train_raw_df

metadata_raw_df = metadata.to_pandas()
metadata_raw_df

# convert timestamp
train_df = train_raw_df.pipe(parse_dt).pipe(handle_dtypes)
train_df

## Distribution per time 

def plot_interaction_by_dayofweek(df, timezone="US/Eastern"):
    plot_df = df.assign(
        day_of_week=df["timestamp"]
        .dt.tz_localize("UTC")
        .dt.tz_convert(timezone)
        .dt.day_name()
    )

    interaction_count_by_day = (
        plot_df.groupby("day_of_week")
        .size()
        .reindex(
            [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday",
            ]
        )
    )

    # Create the plot
    fig = px.bar(
        x=interaction_count_by_day.index,
        y=interaction_count_by_day.values,
        labels={"x": "Day of the Week", "y": "Number of Interactions"},
        title=f"Interaction Count by Day of the Week (assuming {timezone} timezone)",
        text=[f"{val:,.0f}" for val in interaction_count_by_day.values],
        height=500,
        width=700,
    )

    fig.update_layout(xaxis_tickmode="linear")

    fig.show()


timezone = "US/Eastern"  # Assume this time zone because the data is Amazon Data
train_sample_df = train_df.sample(
    args.report_sample_num_rows, replace=False, random_state=args.random_seed
)
plot_interaction_by_dayofweek(train_sample_df, timezone)

def plot_interaction_by_hour(df, timezone="US/Eastern"):
    # Extract the hour from the timestamp
    plot_df = df.assign(
        hour_of_day=df["timestamp"]
        .dt.tz_localize("UTC")
        .dt.tz_convert(timezone)
        .dt.hour
    )

    # Group by the hour of the day and count the number of interactions
    interaction_count_by_hour = (
        plot_df.groupby("hour_of_day").size().reindex(range(24), fill_value=0)
    )

    # Create the plot
    fig = px.bar(
        x=interaction_count_by_hour.index,
        y=interaction_count_by_hour.values,
        labels={"x": "Hour of the Day", "y": "Number of Interactions"},
        title=f"Interaction Count by Hour of the Day (assuming {timezone} timezone)",
        text=[f"{val:,.0f}" for val in interaction_count_by_hour.values],
        height=500,
        width=1200,
    )

    fig.update_layout(xaxis_tickmode="linear")

    fig.show()


timezone = "US/Eastern"  # Assume this time zone because the data is Amazon Data
plot_interaction_by_hour(train_sample_df, timezone)

def plot_interaction_by_month(df, timezone="US/Eastern"):
    # Ensure timestamp is in datetime format
    if not pd.api.types.is_datetime64_any_dtype(df["timestamp"]):
        df["timestamp"] = pd.to_datetime(df["timestamp"])

    # Extract the month from the timestamp (localized)
    plot_df = df.assign(
        month_of_year=df["timestamp"]
        .dt.tz_localize("UTC")
        .dt.tz_convert(timezone)
        .dt.month
    )

    # Group by month and count number of interactions
    interaction_count_by_month = (
        plot_df.groupby("month_of_year").size().reindex(range(1, 13), fill_value=0)
    )

    # Plot
    fig = px.bar(
        x=interaction_count_by_month.index,
        y=interaction_count_by_month.values,
        labels={"x": "Month", "y": "Number of Interactions"},
        title=f"Interaction Count by Month (assuming {timezone} timezone)",
        text=[f"{val:,.0f}" for val in interaction_count_by_month.values],
        height=500,
        width=1200,
    )

    fig.update_layout(
        xaxis_tickmode="linear",
        xaxis_tickvals=interaction_count_by_month.index,
        xaxis_ticktext=[
            pd.to_datetime(str(month), format="%m").strftime("%B")
            for month in interaction_count_by_month.index
        ],
    )

    fig.show()


timezone = "US/Eastern"
plot_interaction_by_month(train_sample_df, timezone)

## Sparsity


def calculate_sparsity(df, user_col="user_id", item_col="parent_asin"):
    return 1 - df.shape[0] / (df[user_col].nunique() * df[item_col].nunique())


print(f"Sparsity: {calculate_sparsity(train_df):,.4%}")
